+++ module/importer/src/Infrastructure/Action/MultiSelectAttributeImportAction.php
@@ -25,7 +25,7 @@ class MultiSelectAttributeImportAction extends AbstractAttributeImportAction
     public function action(ImportMultiSelectAttributeCommand $command): void
     {
         $this->validate($command);
-        $attribute = $this->updateExistingAttribute($command);
+        $attribute = $this->findAttribute(new AttributeCode($command->getCode()));
         if (!$attribute) {
             $attribute = new MultiSelectAttribute(
                 AttributeId::fromKey($command->getCode()),
@@ -35,6 +35,8 @@ class MultiSelectAttributeImportAction extends AbstractAttributeImportAction
                 new TranslatableString($command->getPlaceholder()),
                 new AttributeScope($command->getScope())
             );
+        } else {
+            $this->updateAttribute($command, $attribute);
         }
         $this->processSuccessfulImport($attribute, $command);
     }