+++ module/importer/src/Infrastructure/Action/VariableProductImportAction.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 
 namespace Ergonode\Importer\Infrastructure\Action;
 
+use Ergonode\Product\Domain\Factory\ProductFactoryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\ProductId;
 use Ergonode\Product\Domain\Query\ProductQueryInterface;
 use Ergonode\Product\Domain\ValueObject\Sku;
@@ -46,6 +47,8 @@ class VariableProductImportAction
 
     private ImportProductAttributeBuilder $builder;
 
+    protected ProductFactoryInterface $productFactory;
+
     public function __construct(
         ProductQueryInterface $productQuery,
         ProductRepositoryInterface $productRepository,
@@ -53,7 +56,8 @@ class VariableProductImportAction
         AttributeRepositoryInterface $attributeRepository,
         TemplateQueryInterface $templateQuery,
         CategoryQueryInterface $categoryQuery,
-        ImportProductAttributeBuilder $builder
+        ImportProductAttributeBuilder $builder,
+        ProductFactoryInterface $productFactory
     ) {
         $this->productQuery = $productQuery;
         $this->productRepository = $productRepository;
@@ -62,6 +66,7 @@ class VariableProductImportAction
         $this->templateQuery = $templateQuery;
         $this->categoryQuery = $categoryQuery;
         $this->builder = $builder;
+        $this->productFactory = $productFactory;
     }
 
     /**
@@ -92,7 +97,8 @@ class VariableProductImportAction
 
         if (!$productId) {
             $productId = ProductId::generate();
-            $product = new VariableProduct(
+            $product = $this->productFactory->create(
+                VariableProduct::TYPE,
                 $productId,
                 $sku,
                 $templateId,