+++ module/product/src/Infrastructure/Handler/Create/CreateGroupingProductCommandHandler.php
@@ -11,26 +11,32 @@ namespace Ergonode\Product\Infrastructure\Handler\Create;
 
 use Ergonode\Product\Domain\Entity\GroupingProduct;
 use Ergonode\Product\Domain\Command\Create\CreateGroupingProductCommand;
-use Ergonode\Product\Infrastructure\Handler\AbstractCreateProductHandler;
+use Ergonode\Product\Domain\Factory\ProductFactoryInterface;
+use Ergonode\Product\Domain\Repository\ProductRepositoryInterface;
 
-class CreateGroupingProductCommandHandler extends AbstractCreateProductHandler
+class CreateGroupingProductCommandHandler
 {
+    protected ProductRepositoryInterface $productRepository;
+
+    protected ProductFactoryInterface $productFactory;
+
+    public function __construct(ProductRepositoryInterface $productRepository, ProductFactoryInterface $productFactory)
+    {
+        $this->productRepository = $productRepository;
+        $this->productFactory = $productFactory;
+    }
     /**
      * @throws \Exception
      */
     public function __invoke(CreateGroupingProductCommand $command): void
     {
-        $attributes = $command->getAttributes();
-
-        $attributes = $this->addAudit($attributes);
-        $attributes = $this->addStatusAttribute($attributes);
-
-        $product = new GroupingProduct(
+        $product = $this->productFactory->create(
+            GroupingProduct::TYPE,
             $command->getId(),
             $command->getSku(),
             $command->getTemplateId(),
             $command->getCategories(),
-            $attributes,
+            $command->getAttributes(),
         );
 
         $this->productRepository->save($product);