+++ module/product/src/Infrastructure/Persistence/Projector/DbalProductValueAddedEventProjector.php
@@ -28,5 +28,6 @@ class DbalProductValueAddedEventProjector extends AbstractProductValueProjector
         if (StatusSystemAttribute::CODE !== $code) {
             $this->insertValue($productId, $attributeId, $event->getValue());
         }
+        $this->updateAudit($event->getAggregateId());
     }
 }