+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - feature [#1208](https://github.com/ergonode/backend/issues/1208) Created custom ObjectManager for refresh token and decoupled from Doctrine ORM (piotrkreft)
 - ci [#1212](https://github.com/ergonode/backend/issues/1212) Split CI to test code separately (piotrkreft)
 - performance [#1211](https://github.com/ergonode/backend/issues/1211) Cached user authenticated on workers (piotrkreft)
+- feature [#1218](https://github.com/ergonode/backend/issues/1218) Refactoring of adding status to product  (wiewiurdp)
 - feature [#1174](https://github.com/ergonode/backend/issues/1174) Fixed all allowed Composer dependencies and enabled CI jobs to test against boundary conditions (piotrkreft)
 - bugfix [#1196](https://github.com/ergonode/backend/issues/1196) Decoupled EventSourcing from Core (piotrkreft)
 - refactor [#1199](https://github.com/ergonode/backend/issues/1199) JWT manager to handle stringable user identifier (piotrkreft)